// src/modules/blogger/api/dto/comment-with-additional-info.model.ts
import { LikesModel } from '../../../public/likes/infrastructure/entity/likes.model';
import { ReactionModel } from '../../../../global-model/reaction.model';

export class CommentWithAdditionalInfoModel {
  id: string;
  content: string;
  createdAt: string;
  likesInfo: {
    likesCount: number;
    dislikesCount: number;
    myStatus: string;
  };
  commentatorInfo: {
    userId: string;
    userLogin: string;
  };
  postInfo: {
    id: string;
    title: string;
    blogId: string;
    blogName: string;
  };
}